+++ app/src/main/java/protect/card_locker/DBHelper.java
@@ -6,7 +6,6 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteException;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.graphics.Bitmap;
 
 import com.google.zxing.BarcodeFormat;
 
@@ -275,7 +274,7 @@ public class DBHelper extends SQLiteOpenHelper
         return newId;
     }
 
-    public boolean insertLoyaltyCard(final SQLiteDatabase db, final String store,
+    public long insertLoyaltyCard(final SQLiteDatabase db, final String store,
                                      final String note, final Date expiry, final BigDecimal balance,
                                      final Currency balanceType, final String cardId,
                                      final String barcodeId, final BarcodeFormat barcodeType,
@@ -293,10 +292,10 @@ public class DBHelper extends SQLiteOpenHelper
         contentValues.put(LoyaltyCardDbIds.HEADER_COLOR, headerColor);
         contentValues.put(LoyaltyCardDbIds.STAR_STATUS,starStatus);
         final long newId = db.insert(LoyaltyCardDbIds.TABLE, null, contentValues);
-        return (newId != -1);
+        return newId;
     }
 
-    public boolean insertLoyaltyCard(final SQLiteDatabase db, final int id, final String store,
+    public long insertLoyaltyCard(final SQLiteDatabase db, final int id, final String store,
                                      final String note, final Date expiry, final BigDecimal balance,
                                      final Currency balanceType, final String cardId,
                                      final String barcodeId, final BarcodeFormat barcodeType,
@@ -315,7 +314,7 @@ public class DBHelper extends SQLiteOpenHelper
         contentValues.put(LoyaltyCardDbIds.HEADER_COLOR, headerColor);
         contentValues.put(LoyaltyCardDbIds.STAR_STATUS,starStatus);
         final long newId = db.insert(LoyaltyCardDbIds.TABLE, null, contentValues);
-        return (newId != -1);
+        return newId;
     }
 
     public boolean updateLoyaltyCard(final int id, final String store, final String note,