+++ app/src/main/java/protect/card_locker/LoyaltyCard.java
@@ -1,10 +1,6 @@
 package protect.card_locker;
 
 import android.database.Cursor;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-
-import java.sql.Blob;
 
 import com.google.zxing.BarcodeFormat;