+++ app/src/main/java/protect/card_locker/importexport/DataFormat.java
@@ -1,9 +1,10 @@
-package protect.card_locker;
+package protect.card_locker.importexport;
 
 public enum DataFormat
 {
     Catima,
     Fidme,
+    Stocard,
     VoucherVault
     ;
 }