+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changes:
 
 - Add UPC-E support
 - Support adding a front and back photo to each card
+- Support importing password-protected zip files
+- Support importing from Stocard (Beta)
 - Fix useless whitespace in notes from Fidme import
 
 ## v1.14.1 (2021-06-14)